tests/wpt/css-tests/css-transforms-1_dev/html/transform_translate_zero.htm

<!DOCTYPE html>
<html><head>
    <title>CSS Transforms API Test: transform translate</title>
    <link href="http://www.intel.com" rel="author" title="Intel">
    <link href="http://www.w3.org/TR/css-transforms-1/#transform-property" rel="help">
    <link href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions" rel="help">
    <meta content="dom" name="flags">
    <meta content="Check if transform supports translate(0, 0)" name="assert"> 
    <script src="/resources/testharness.js" type="text/javascript"></script>
    <script src="/resources/testharnessreport.js" type="text/javascript"></script>    
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
    test(function() {
        document.getElementById("test").style.transform = "translate(0, 0)";
        var value = document.getElementById("test").style.getPropertyCSSValue("transform").cssText;
        assert_equals(value, "translate(0px, 0px)")
    }, "transform_translate_0_0");
    </script>    


</body></html>
